REARRANGED Layout into a folder
