+++ frontend/src/components/Layout/components/Footer.jsx
@@ -1,0 +1,11 @@
+function Footer() {
+    return (
+      <footer className="relative z-20 bg-transparent py-4 sm:py-6 text-center footer-animation">
+        <p className="text-sm sm:text-base text-blue-300/80">
+          &copy; 2025 Investigators. Todos los derechos reservados.
+        </p>
+      </footer>
+    );
+  }
+  
+  export default Footer;
