+++ frontend/src/components/Layout/components/MenuOverlay.jsx
@@ -1,0 +1,10 @@
+function MenuOverlay({ isOpen, onClose }) {
+    return (
+      <div
+        className={`menu-overlay ${isOpen ? "open" : ""}`}
+        onClick={onClose}
+      ></div>
+    );
+  }
+  
+  export default MenuOverlay;
